+++ src/localisation/chineseSimplified.ts
@@ -105,7 +105,7 @@ const CHINESE_SIMPLIFIED: Translations = {
   [Phrase.SeparateBufferFolderLabel]: '独立的缓冲文件夹',
   [Phrase.BufferFolderLabel]: '缓冲文件夹',
   [Phrase.MaxDiskStorageLabel]: '磁盘存储上限 (GB)',
-  [Phrase.WindowsSettingsLabel]: 'Windows 设置',
+  [Phrase.AppSettingsLabel]: '应用 设置',
   [Phrase.RunOnStartupLabel]: '开机自动启动',
   [Phrase.StartMinimizedLabel]: '启动时最小化',
   [Phrase.MinimizeOnQuitLabel]: '关闭时最小化',
@@ -413,6 +413,8 @@ const CHINESE_SIMPLIFIED: Translations = {
   [Phrase.SelectProcess]: "选择一个应用程序",
   [Phrase.AudioProcessDevicesDescription]: "除了任何选定的扬声器和麦克风之外，还可以从中捕获音频的应用程序。",
   [Phrase.ProcessVolumeDescription]: "录音中应用程序的音量，从0到1。",
+  [Phrase.HideEmptyCategoriesLabel]: "隐藏空类别",
+  [Phrase.HideEmptyCategoriesDescription]: "隐藏侧面菜单中没有视频的类别。",
 };
 
 export default CHINESE_SIMPLIFIED;