+++ src/localisation/english.ts
@@ -105,7 +105,7 @@ const ENGLISH: Translations = {
   [Phrase.SeparateBufferFolderLabel]: 'Separate Buffer Folder',
   [Phrase.BufferFolderLabel]: 'Buffer Folder',
   [Phrase.MaxDiskStorageLabel]: 'Max Disk Storage (GB)',
-  [Phrase.WindowsSettingsLabel]: 'Windows Settings',
+  [Phrase.AppSettingsLabel]: 'Application Settings',
   [Phrase.RunOnStartupLabel]: 'Run on Startup',
   [Phrase.StartMinimizedLabel]: 'Start Minimized',
   [Phrase.MinimizeOnQuitLabel]: 'Minimize on Quit',
@@ -413,6 +413,8 @@ const ENGLISH: Translations = {
   [Phrase.SelectProcess]: "Select an application",
   [Phrase.AudioProcessDevicesDescription]: "Applications to capture audio from, in addition to any selected speakers and microphones.",
   [Phrase.ProcessVolumeDescription]: "The volume of the application in the recording, from 0 to 1.",
+  [Phrase.HideEmptyCategoriesLabel]: "Hide Empty Categories",
+  [Phrase.HideEmptyCategoriesDescription]: "Hides categories in the side menu that have no videos in them.",
 };
 
 export default ENGLISH;